network-tests: fix types imports after refactor
